共有ボタンを車両設定画面に移動
- トップバーの車両切り替えから共有ボタンを削除
- 車両設定画面に共有相手を追加するボタンを追加
